feat(env): rename to space_KOZ
